requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyyaml==6.0.1
crewai>=0.159.0,<1.0.0
crewai-tools>=0.1.0
python-dotenv==1.0.0
pdfplumber>=0.11.4
tabulate==0.9.0
pypdf2==3.0.1
markdown2==2.4.11
yfinance==0.2.33
pandas>=2.1.4
numpy>=1.26.2